1.修改bean.

+++ business-center/com-center/com-service/business-basic/src/main/java/com/szmsd/bas/controller/BasMainController.java
@@ -1,8 +1,8 @@
 package com.szmsd.bas.controller;
 
 import com.alibaba.fastjson.JSONObject;
-import com.szmsd.bas.domain.BasMain;
-import com.szmsd.bas.domain.BasSub;
+import com.szmsd.bas.api.domain.BasMain;
+import com.szmsd.bas.api.domain.BasSub;
 import com.szmsd.bas.service.IBasMainService;
 import com.szmsd.bas.service.IBasSubService;
 import com.szmsd.common.core.domain.R;
@@ -105,7 +105,7 @@ public class BasMainController extends BaseController {
             }
             basMain.setCreateTime(new Date());
             basMainService.insertBasMain(basMain);
-            int t = basMain.getId();
+            Long t = basMain.getId();
             String s = String.format("%03d", t);
             basMain.setMainCode(s);
             basMainService.updateBasMain(basMain);
